Include toolbar in entity table component
